IngredientService fin. (TDD)

+++ src/main/java/cp/demo/domain/entity/MyIngredientEntity.java
@@ -27,7 +27,7 @@ public class MyIngredientEntity {
     private String ingredName;
 
     @Column(name = "capacity", columnDefinition = "INT")
-    private Long capacity;
+    private int capacity;
 
     @Column(name = "unit", columnDefinition = "varchar(256)")
     private String unit;
